Refactor Oracle client into partials
