+++ DbaClientX.Oracle/Oracle.Transactions.cs
@@ -1,0 +1,211 @@
+using System;
+using System.Data;
+using System.Threading;
+using System.Threading.Tasks;
+using Oracle.ManagedDataAccess.Client;
+
+namespace DBAClientX;
+
+public partial class Oracle
+{
+    /// <summary>
+    /// Begins a transaction using the Oracle connection derived from the provided credentials.
+    /// </summary>
+    public virtual void BeginTransaction(string host, string serviceName, string username, string password)
+        => BeginTransaction(host, serviceName, username, password, IsolationLevel.ReadCommitted);
+
+    /// <summary>
+    /// Begins a transaction using the Oracle connection derived from the provided credentials and isolation level.
+    /// </summary>
+    public virtual void BeginTransaction(string host, string serviceName, string username, string password, IsolationLevel isolationLevel)
+    {
+        lock (_syncRoot)
+        {
+            if (_transaction != null)
+            {
+                throw new DbaTransactionException("Transaction already started.");
+            }
+
+            var connectionString = BuildConnectionString(host, serviceName, username, password);
+            _transactionConnection = new OracleConnection(connectionString);
+            _transactionConnection.Open();
+            _transaction = _transactionConnection.BeginTransaction(isolationLevel);
+        }
+    }
+
+    /// <summary>
+    /// Asynchronously begins a transaction using the Oracle connection derived from the provided credentials.
+    /// </summary>
+    public virtual Task BeginTransactionAsync(string host, string serviceName, string username, string password, CancellationToken cancellationToken = default)
+        => BeginTransactionAsync(host, serviceName, username, password, IsolationLevel.ReadCommitted, cancellationToken);
+
+    /// <summary>
+    /// Asynchronously begins a transaction using the Oracle connection derived from the provided credentials and isolation level.
+    /// </summary>
+    public virtual async Task BeginTransactionAsync(string host, string serviceName, string username, string password, IsolationLevel isolationLevel, CancellationToken cancellationToken = default)
+    {
+        lock (_syncRoot)
+        {
+            if (_transaction != null)
+            {
+                throw new DbaTransactionException("Transaction already started.");
+            }
+        }
+
+        var connectionString = BuildConnectionString(host, serviceName, username, password);
+        var connection = new OracleConnection(connectionString);
+        await connection.OpenAsync(cancellationToken).ConfigureAwait(false);
+#if NETSTANDARD2_1_OR_GREATER || NETCOREAPP3_0_OR_GREATER || NET5_0_OR_GREATER
+        var transaction = (OracleTransaction)await connection.BeginTransactionAsync(isolationLevel, cancellationToken).ConfigureAwait(false);
+#else
+        var transaction = connection.BeginTransaction(isolationLevel);
+#endif
+        lock (_syncRoot)
+        {
+            if (_transaction != null)
+            {
+                transaction.Dispose();
+                connection.Dispose();
+                throw new DbaTransactionException("Transaction already started.");
+            }
+
+            _transactionConnection = connection;
+            _transaction = transaction;
+        }
+    }
+
+    /// <summary>
+    /// Commits the active transaction.
+    /// </summary>
+    public virtual void Commit()
+    {
+        lock (_syncRoot)
+        {
+            if (_transaction == null)
+            {
+                throw new DbaTransactionException("No active transaction.");
+            }
+
+            _transaction.Commit();
+            DisposeTransactionLocked();
+        }
+    }
+
+    /// <summary>
+    /// Asynchronously commits the active transaction.
+    /// </summary>
+    public virtual async Task CommitAsync(CancellationToken cancellationToken = default)
+    {
+        OracleTransaction? tx;
+        OracleConnection? conn;
+        lock (_syncRoot)
+        {
+            if (_transaction == null)
+            {
+                throw new DbaTransactionException("No active transaction.");
+            }
+
+            tx = _transaction;
+            conn = _transactionConnection;
+            _transaction = null;
+            _transactionConnection = null;
+        }
+
+        try
+        {
+#if NETSTANDARD2_1_OR_GREATER || NETCOREAPP3_0_OR_GREATER || NET5_0_OR_GREATER
+            await tx!.CommitAsync(cancellationToken).ConfigureAwait(false);
+#else
+            tx!.Commit();
+#endif
+        }
+        finally
+        {
+            tx!.Dispose();
+            conn?.Dispose();
+        }
+    }
+
+    /// <summary>
+    /// Rolls back the active transaction.
+    /// </summary>
+    public virtual void Rollback()
+    {
+        lock (_syncRoot)
+        {
+            if (_transaction == null)
+            {
+                throw new DbaTransactionException("No active transaction.");
+            }
+
+            _transaction.Rollback();
+            DisposeTransactionLocked();
+        }
+    }
+
+    /// <summary>
+    /// Asynchronously rolls back the active transaction.
+    /// </summary>
+    public virtual async Task RollbackAsync(CancellationToken cancellationToken = default)
+    {
+        OracleTransaction? tx;
+        OracleConnection? conn;
+        lock (_syncRoot)
+        {
+            if (_transaction == null)
+            {
+                throw new DbaTransactionException("No active transaction.");
+            }
+
+            tx = _transaction;
+            conn = _transactionConnection;
+            _transaction = null;
+            _transactionConnection = null;
+        }
+
+        try
+        {
+#if NETSTANDARD2_1_OR_GREATER || NETCOREAPP3_0_OR_GREATER || NET5_0_OR_GREATER
+            await tx!.RollbackAsync(cancellationToken).ConfigureAwait(false);
+#else
+            tx!.Rollback();
+#endif
+        }
+        finally
+        {
+            tx!.Dispose();
+            conn?.Dispose();
+        }
+    }
+
+    private void DisposeTransaction()
+    {
+        lock (_syncRoot)
+        {
+            DisposeTransactionLocked();
+        }
+    }
+
+    private void DisposeTransactionLocked()
+    {
+        _transaction?.Dispose();
+        _transaction = null;
+        _transactionConnection?.Dispose();
+        _transactionConnection = null;
+    }
+
+    /// <inheritdoc />
+    protected override bool IsTransient(Exception ex) =>
+        ex is OracleException oex && (oex.Number == 12541 || oex.Number == 12545 || oex.Number == 1089 || oex.Number == 3113 || oex.Number == 3114);
+
+    /// <inheritdoc />
+    protected override void Dispose(bool disposing)
+    {
+        if (disposing)
+        {
+            DisposeTransaction();
+        }
+
+        base.Dispose(disposing);
+    }
+}
